web-testing/requirements.txt
playwright
google-generativeai>=0.3.0
